pyproject.toml
[tool.poetry]
name = "withbuy-backend"
version = "0.1.0"
description = ""
authors = ["cheeseNA <kurocat2000@gmail.com>"]
readme = "README.md"
packages = [{include = "withbuy_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.1"
motor = "^3.1.1"
pydantic = {extras = ["email"], version = "^1.10.5"}


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
